finish signup and split it into components
